done lab 9

+++ lab_8.py
@@ -29,7 +29,7 @@ def task_5(a):
 
 
 def main():
-    task_1(float(input()))
+    task_1(int(input()))
     task_2(float(input()), float(input()))
     task_3(float(input()), float(input()))
     task_4(int(input()))
